Fix "argument of type 'NoneType' is not iterable".
It appears that in some cases a model event will occur that doesn't have
a snapshot or a parent. I haven't been able to isolate the case where
this occurs, but the safe thing to do is ignore the update and catch any
resulting drift in the period full remodel.

Fixes ZEN-14856.